feat: :sparkles: Propsの型設定
