+++ 05_typescript_v3/src/main.ts
@@ -1,6 +1,7 @@
 import { createApp } from 'vue';
 // import App from './App/App01.vue'
 // import App from './App/App02.vue'
+// import App from './App/App03.vue'
 import App from './App.vue';
 
 import './assets/main.css';
